# docs/swagger.yaml
definitions:
...
  models.Sekolah:
    properties:
      created_at:
        type: string
      id:
        type: integer
      kode_verifikasi_pendaftaran_pendidik:
        type: string
      kode_verifikasi_pendaftaran_siswa:
        type: string
      nama:
        type: string
      pendaftaran_pendidik:
        type: string
      pendaftaran_siswa:
        type: string
      status_menu_pendidik:
        type: string
      status_menu_siswa:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  sekolahController.UpdateInputKodeVerifikasiPendidik:
    properties:
      kode_verifikasi_pendaftaran_pendidik:
        type: string
    type: object
  sekolahController.UpdateInputKodeVerifikasiSiswa:
    properties:
      kode_verifikasi_pendaftaran_siswa:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /sekolah/{id}:
    delete:
      description: Menghapus semua data sekolah berdasarkan id. mulai dari akun, .
      parameters:
      - description: Sekolah id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerToken: []
      summary: Menghapus data sekolah.
      tags:
      - Siswa
    get:
      description: Mendapatkan data sekolah berdasarkan id.
      parameters:
      - description: Sekolah id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sekolah'
      summary: Mendapatkan data sekolah.
      tags:
      - Sekolah
  /sekolah/{id}/aktivasi-menu-pendidik:
    patch:
      description: Aktivasi Menu pengisian/update biodata siswa Siswa berdasarkan
        sekolah id.
      parameters:
      - description: Sekolah id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sekolah'
      security:
      - BearerToken: []
      summary: Aktivasi Menu pengisian/update biodata siswa Siswa berdasarkan sekolah
        id.
      tags:
      - Sekolah
  /sekolah/{id}/aktivasi-menu-siswa:
    patch:
      description: Aktivasi Menu pengisian/update biodata siswa Siswa berdasarkan
        sekolah id.
      parameters:
      - description: Sekolah id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sekolah'
      security:
      - BearerToken: []
      summary: Aktivasi Menu pengisian/update biodata siswa Siswa berdasarkan sekolah
        id.
      tags:
      - Sekolah
...
  /sekolah/{id}/buka-pendaftaran-pendidik:
    patch:
      description: Buka pendafatran pendidik berdasarkan sekolah id.
      parameters:
      - description: Sekolah id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sekolah'
      security:
      - BearerToken: []
      summary: Buka pendafatran pendidik berdasarkan sekolah id.
      tags:
      - Sekolah
  /sekolah/{id}/buka-pendaftaran-siswa:
    patch:
      description: Buka pendafatran siswa berdasarkan sekolah id.
      parameters:
      - description: Sekolah id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sekolah'
      security:
      - BearerToken: []
      summary: Buka pendafatran siswa berdasarkan sekolah id.
      tags:
      - Sekolah
  /sekolah/{id}/kode-verifikasi-pendaftaran-pendidik:
    patch:
      description: memperbaharui kode verifikasi pendaftaran pendidik berdasarkan
        id sekolah.
      parameters:
      - description: Tahun ajaran id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update kode verifikasi pendaftaran pendidik
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/sekolahController.UpdateInputKodeVerifikasiPendidik'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sekolah'
      security:
      - BearerToken: []
      summary: Memperbarui kode verifikasi pendaftaran pendidik.
      tags:
      - Sekolah
  /sekolah/{id}/kode-verifikasi-pendaftaran-siswa:
    patch:
      description: memperbaharui kode verifikasi pendaftaran siswa berdasarkan id
        sekolah.
      parameters:
      - description: Tahun ajaran id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to update kode verifikasi pendaftaran siswa
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/sekolahController.UpdateInputKodeVerifikasiSiswa'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sekolah'
      security:
      - BearerToken: []
      summary: Memperbarui kode verifikasi pendaftaran siswa.
      tags:
      - Sekolah
  /sekolah/{id}/non-aktif-menu-pendidik:
    patch:
      description: non aktif Menu pengisian/update biodata siswa Siswa berdasarkan
        sekolah id.
      parameters:
      - description: Sekolah id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sekolah'
      security:
      - BearerToken: []
      summary: non aktif Menu pengisian/update biodata siswa Siswa berdasarkan sekolah
        id.
      tags:
      - Sekolah
  /sekolah/{id}/non-aktif-menu-siswa:
    patch:
      description: Non aktif Menu pengisian/update biodata siswa Siswa berdasarkan
        sekolah id.
      parameters:
      - description: Sekolah id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sekolah'
      security:
      - BearerToken: []
      summary: Non aktif Menu pengisian/update biodata siswa Siswa berdasarkan sekolah
        id.
      tags:
      - Sekolah
...
  /sekolah/{id}/tutup-pendaftaran-pendidik:
    patch:
      description: Tutup pendafatran pendidik berdasarkan sekolah id.
      parameters:
      - description: Sekolah id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sekolah'
      security:
      - BearerToken: []
      summary: Tutup pendafatran pendidik berdasarkan sekolah id.
      tags:
      - Sekolah
  /sekolah/{id}/tutup-pendaftaran-siswa:
    patch:
      description: Tutup pendafatran siswa berdasarkan sekolah id.
      parameters:
      - description: Sekolah id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sekolah'
      security:
      - BearerToken: []
      summary: Tutup pendafatran siswa berdasarkan sekolah id.
      tags:
      - Sekolah
...
  /sekolah/user/{id}:
    get:
      description: Mendapatkan data sekolah berdasarkan user ID.
      parameters:
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sekolah'
      summary: Mendapatkan data sekolah berdasarkan user ID.
      tags:
      - Sekolah
...
swagger: "2.0"